README: Update for alternative acme-client
acme-client-portable is no longer maintained by its previous maintainer
as acme-client is now a part of OpenBSD and he is only concentrating on
that[0].

However someone has continued maintaining a fork[1] of this and there
have been various changes to the program since. E.g the program is now
mostly driven via a config file.

Update the README to reflect this new state of things.

[0]: https://kristaps.bsd.lv/acme-client/
[1]: https://github.com/graywolf/acme-client-portable

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>