Create our own listen socket
Rather than have libmicrohttpd create the listen socket, create it
ourselves and pass it to it.

This will allow to start the http server later than it currently does,
removing a window where the server is up and accepting connections
_before_ all the sandboxing has been done.

Seeing as we can only pass a single listen socket into libmicrohttpd we
just create an IPv6 listen socket but with the IPV6_V6ONLY socket option
set to 0, so the socket will also accept v4 connections using the
v4-mapped-on-v6 address type (i.e. ::ffff:xxx.xxx.xxx.xxx).

That is the same behaviour as we had before with the MHD_USE_IPv6 |
MHD_USE_DUAL_STACK flags.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>