Start the web server _after_ we have switched user
This removes a window whereby we are accepting http connections while
running as root.

We now start libmicrohttpd just before initialising seccomp(2). We could
start it _after_ seccomp(2) but then we need to allow a bunch more
syscalls just to start it. So I think it's better we have a small window
before starting seccomp(2) but with a more restrictive set of allowed
syscalls.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>